refactor: extract subcomponent Selection into new file
