+++ src/views/components/Selection/ISelection.tsx
@@ -1,0 +1,7 @@
+import { Game, Genre } from '@data/types'
+
+
+export interface ISelection {
+	games: Game[]
+	genres: Genre[]
+}
