+++ src/views/components/index.ts
@@ -5,3 +5,4 @@ export * from './GameCard/GameCard'
 export * from './GamePanel/GamePanel'
 export * from './ImageCard/ImageCard'
 export * from './Gallery/Gallery'
+export * from './Selection/Selection'
