+++ src/views/routes/Home/Home.tsx
@@ -1,9 +1,8 @@
 import { RootContext }                     from '@data/context'
 import { Game, Genre, Recommended }        from '@data/types'
 import { db }                              from '@src/main.tsx'
-import { Gallery, GameCard, GamePanel }    from '@views/components'
+import { Gallery, GamePanel, Selection }   from '@views/components'
 import { useContext, useEffect, useState } from 'react'
-import { Link }                            from 'react-router-dom'
 import style                               from './Home.module.scss'
 
 
@@ -79,43 +78,5 @@ export function Home() {
 			</article>
 		</main>
 	)
-}
-
-interface ISelection {
-	games: Game[]
-	genres: Genre[]
-}
-
-function Selection({ games, genres }: ISelection) {
-	const gameList = games.map(g =>
-								   (<li key={g.id}>
-									   <GameCard game={g}/>
-								   </li>))
-
-	const genreList = genres.map(g => {
-		const key = `selection-${typeof (g)}-${g.id}`
-
-		return (
-			<li key={key}>
-				<button className={style.Option}>{g.name}</button>
-			</li>
-		)
-	})
-
-	return (
-		<article className={style.Selection}>
-			<h3>Navigate by genre</h3>
-
-			<menu className={style.Menu}>
-				{genreList}
-			</menu>
-
-			<ul className={style.GameList}>
-				{gameList}
-				<li>
-					<Link to="/search">see all</Link>
-				</li>
-			</ul>
-		</article>)
 }
